backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
langchain==0.2.0
langchain-groq==0.1.5
redis==5.0.4
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
